update kernel
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@61827 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/linux/universal/linux-6.12/tools/include/nolibc/stdlib.h
@@ -274,7 +274,7 @@ int itoa_r(long in, char *buffer)
 	int len = 0;
 
 	if (in < 0) {
-		in = -(unsigned long)in;
+		in = -in;
 		*(ptr++) = '-';
 		len++;
 	}
@@ -410,7 +410,7 @@ int i64toa_r(int64_t in, char *buffer)
 	int len = 0;
 
 	if (in < 0) {
-		in = -(uint64_t)in;
+		in = -in;
 		*(ptr++) = '-';
 		len++;
 	}
